🐛🤖 Fix EntitySelector Fast Refresh invalidation
Move EXTERNAL_SORT_INDICATOR_DEFINITIONS out of the React component module and into EntitySelectorConstants. Vite Fast Refresh expects component modules to export only components, so the previous non-component export caused HMR invalidation warnings while editing EntitySelector.
